expose some methods from insights class
