+++ scopes/explorer/insights/insights.cmd.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
 import { Command, CommandOptions } from '@teambit/cli';
 import { InsightResult } from './insight';
-import { InsightManager, RunInsightOptions } from './insight-manager';
+import { InsightsMain } from './insights.main.runtime';
 
 export default class InsightsCmd implements Command {
   name = 'insights [...names]';
@@ -12,37 +12,23 @@ export default class InsightsCmd implements Command {
     ['l', 'list', 'list all insights'],
     ['j', 'json', 'return the insights in json format'],
   ] as CommandOptions;
-  insightManager: InsightManager;
-  constructor(insightManager: InsightManager) {
-    this.insightManager = insightManager;
-  }
+  constructor(private insights: InsightsMain) {}
 
   async report(names: [string[]], options: { list: boolean }): Promise<string> {
     if (options.list) {
       const results = await this.json(names, options);
       return JSON.stringify(results, null, 2);
     }
-    const results = await this.runInsights(names, { renderData: true });
+    const results = await this.insights.runInsights(names, { renderData: true });
     return template(results);
   }
 
   async json(names: [string[]], { list }: { list: boolean }) {
     if (list) {
-      const results = this.insightManager.listInsights();
+      const results = this.insights.listInsights();
       return results;
     }
-    return this.runInsights(names, { renderData: false });
-  }
-
-  private async runInsights([names]: [string[]], opts: RunInsightOptions) {
-    if (names) {
-      let results: InsightResult[] = [];
-      const namesArr = typeof names === 'string' ? [names] : names;
-      results = await this.insightManager.run(namesArr, opts);
-      return results;
-    }
-    const results = await this.insightManager.runAll(opts);
-    return results;
+    return this.insights.runInsights(names, { renderData: false });
   }
 }
